# Performance improvement getClassNames