--- a/src/utils.js
+++ b/src/utils.js
@@ -27,3 +27,19 @@ function getCSS(styleSheet) {
 }
 
 module.exports.getCSS = getCSS
+
+function getClassNames (node, classNames) {
+  if (node.children && node.children.reduce) {
+    classNames = node.children.reduce((acc, child) => (
+      acc.concat(getClassNames(child, acc))
+    ), classNames)
+  }
+
+  if (node.props && node.props.className) {
+    return classNames.concat(node.props.className.split(' '))
+  }
+
+  return classNames
+}
+
+module.exports.getClassNames = getClassNames