It might be hard to tell in the diff, this is the line that changed:
`acc.concat(getClassNames(child, []))`

I thought of another solution that uses `push` instead of `concat` to save memory, not sure if it's necessary though... so I'll just open this PR for meow

EDIT:

This was happening until I made the change in this PR

```
 FAIL  ./storyshots.test.js (10.358s)
  ● Storyshots › ReactSwagger › default

    RangeError: Invalid array length

      at Object.test (node_modules/@storybook/addon-storyshots/dist/test-bodies.js:22:18)
      at Object.<anonymous> (node_modules/@storybook/addon-storyshots/dist/index.js:147:25)
      at Promise.resolve.then.el (node_modules/p-map/index.js:42:16)
      at process._tickCallback (internal/process/next_tick.js:109:7)

  ● Storyshots › ReactSwagger › no title

    RangeError: Invalid array length

      at Object.test (node_modules/@storybook/addon-storyshots/dist/test-bodies.js:22:18)
      at Object.<anonymous> (node_modules/@storybook/addon-storyshots/dist/index.js:147:25)
      at Promise.resolve.then.el (node_modules/p-map/index.js:42:16)
      at process._tickCallback (internal/process/next_tick.js:109:7)

  Storyshots
    ReactSwagger
      ✕ default (6864ms)
      ✕ no title (2718ms)

Test Suites: 1 failed, 1 total
Tests:       2 failed, 2 total
Snapshots:   0 total
Time:        10.423s, estimated 11s
Ran all test suites.

Watch Usage: Press w to show more.
```