@MicheleBertoli 2017-06-22
Interesting, thanks @wongterrencew.
This might be a regression of a change I recently introduced.
Do you mind providing an example of the component that makes the tests fail?