@MicheleBertoli 2017-06-22
Hello @wongterrencew, I managed to repro the issue. 
Thank you very much!