@MicheleBertoli 2017-06-22
Also, since we were always passing `[]` I realised we didn't even need to pass down the array anymore so I decided to go with the `push` solution (I'm pretty sure it's similar to what you had in mind).
v3.1.3 is out, thank you very much @wongterrencew